more frontend changes
